requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.8.5
drf-extensions==0.7.1